Had to change transformation logic to allow for a wider canvas
While the game board is square, the canvas should be bigger to allow the
player to pick different pieces. Making the canvas wider broke all the
transformation code, since it was based on assumptions that the canvas was
square. Fixed all the code so now it works regardless of the canvas width.

	modified:   blockus.js
	modified:   index.html