Fixed placing bug and almost done game
The bug where you could incorrectly place a piece was fixed. It turns out
that when you would select one piece and then select another, the rotation
of first selected piece wasn't getting reset. This means the piece was
displayed one way but different in the model.

Created a function which checks all available hypothetical moves for each
player, to determine if they have run out of moves. It needs to be tested,
but if it is working then the game can end properly.

Made minor fixes to function/variable names to be consistent (only uses
the term 'square' and doesn't sometimes have 'block'

Worked on #2, #3
Fixes #7

	modified:   blockus.js