Hopefully final attempt at fixing placement bug
Now the issue was that the reset function wasn't resetting properly, since
it was rotating then flipping, instead of vice versa. Now hopefully a
piece is reset to the way it should be after hypothetically arranging it.

Also rearranged the code for when to check if a player has moves left. If
only checked after they make a move, then if another player makes it so
they can no longer play, it wasn't detected. Now the check is before they
start, to see if any recent moves kicked them out of the game.

Also reports who won.

	modified:   blockus.js