Fixed bug when deselecting piece, added debug code
When deselecting a piece, the game would crash. This was because the game
was "selecting" a non existant piece. This has been fixed.

Some times when the game reaches the end, it seems like there's no more
moves for a player, but they can still play. Added debug code to see if
there really are available moves or not

	modified:   blockus.js